Move GenerateBillRunService
